refactor(server): move folder
