+++ server/src/app/http/routes.ts
@@ -1,9 +1,9 @@
 import express from 'express'
 
-import registerClassesRoutes from '../../modules/classes/routes'
-import registerClassSchedulesRoutes from '../../modules/classSchedules/routes'
-import connectionsRoutes from '../../modules/connections/routes'
-import usersRoutes from '../../modules/users/routes'
+import registerClassesRoutes from '../modules/classes/routes'
+import registerClassSchedulesRoutes from '../modules/classSchedules/routes'
+import connectionsRoutes from '../modules/connections/routes'
+import usersRoutes from '../modules/users/routes'
 
 function register(app: express.Application) {
   const router = express.Router()
