+++ server/src/app/modules/classes/routes.ts
@@ -1,6 +1,6 @@
 import express from 'express'
 
-import ClassController from '../../controllers/ClassController'
+import ClassController from '../../../controllers/ClassController'
 
 function register(prefix: string, appRouter: express.Router) {
   const router = express.Router()
